RPG Common: Started base of the common code
This common code works on both the server and client. It includes things
like map generation and game logic. It completely excludes rendering
because that is the job of the client.

The abstract Game object allows for a headless game loop, for use on the
server side. Map2D creates a 2D map abstraction. The World object
exports tile management.

Since these are pieces of common code, their implementation get
synchronized with the client code.