RPG Common: Proper collision resolution
Goodness gracious, this is a lot of code.
Currently, it is perfect :D
This also fixes various logic errors in collision.dart and entity.dart,
notably how collisions actually calculate their resolutions.
This also gives TestEntities IDs for ease of debugging.
This also gives every CollisionObject an anchored parameter.
This also lets getCollisions filter collisions.
This also adds the opposite getter to CollisionResponse, which will
create an opposite version of the current response.